api/requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.7.0
attrs==24.3.0
certifi==2024.12.14
charset-normalizer==3.4.0
click==8.1.7
eval-type-backport==0.2.0
fastapi==0.115.6
filelock==3.16.1
frozenlist==1.5.0
h11==0.14.0
idna==3.10
lxml==5.3.0
markdown-it-py==3.0.0
mdurl==0.1.2
multidict==6.1.0
numpy==2.2.0
pillow==10.4.0
propcache==0.2.1
pyarrow==18.1.0
pydantic==2.10.4
pydantic-core==2.27.2
pygments==2.18.0
pymupdf==1.25.1
python-dotenv==1.0.1
python-pptx==1.0.2
regex==2024.11.6
requests==2.32.3
rich==13.9.4
shellingham==1.5.4
sniffio==1.3.1
starlette==0.41.3
tabulate==0.9.0
tiktoken==0.8.0
together==1.3.10
tqdm==4.67.1
typer==0.15.1
typing-extensions==4.12.2
urllib3==2.2.3
uvicorn==0.34.0
xlsxwriter==3.2.0
yarl==1.18.3
